ChatPageTemplate done, testing, localPage testing
